+++ src/read_data_sphNG.f90
@@ -795,11 +795,7 @@ subroutine extract_variables_from_header(tags,realarr,nreals,iverbose,debug,&
        massoftype(itype)  = massoftypei(i)
     enddo
     npartoftype(itypemap_sink_phantom) = nptmasstot  ! sink particles
-    if (nblocks > 1) then
-       print "(a)",' setting ngas=npart for MPI code '
-       npartoftype(1)  = npart
-       npartoftype(2:) = 0
-    endif
+    if (debug) print*,'DEBUG: npart of type sink ',itypemap_sink_phantom,' = ',nptmasstot
     !
     !--if Phantom calculation uses the binary potential
     !  then read this as two point mass particles
@@ -1018,10 +1014,10 @@ subroutine guess_labels(ncolumns,iamvec,label,labelvec,istartmhd, &
  endif
 end subroutine guess_labels
 
-integer function assign_column(tag,iarr,ipos,ikind,imaxcolumnread,idustarr) result(icolumn)
+integer function assign_column(tag,iarr,ipos,ikind,imaxcolumnread,idustarr,ncolstep) result(icolumn)
  use labels, only:ih,irho,ix,ipmass
  character(len=lentag), intent(in) :: tag
- integer,               intent(in) :: iarr,ipos,ikind
+ integer,               intent(in) :: iarr,ipos,ikind,ncolstep
  integer,               intent(inout) :: imaxcolumnread
  integer,               intent(inout) :: idustarr
 
@@ -1062,6 +1058,10 @@ integer function assign_column(tag,iarr,ipos,ikind,imaxcolumnread,idustarr) resu
        icolumn = nhydroarrays + ndustarrays + 3
     case default
        icolumn = max(nhydroarrays + ndustarrays + nmhdarrays + 1,imaxcolumnread + 1)
+       if (icolumn > ncolstep) then
+          ! check for dustfrac not being present
+          if (idustarr == 0 .and. ndustarrays > 0) icolumn = nhydroarrays + 1
+       endif
        if (iarr==1) then
           if (ikind==4) then  ! real*4 array
              istart_extra_real4 = min(istart_extra_real4,icolumn)
@@ -1120,6 +1120,11 @@ integer function assign_column(tag,iarr,ipos,ikind,imaxcolumnread,idustarr) resu
        icolumn = imaxcolumnread + 1
     endif
  endif
+ if (icolumn < 1 .or. icolumn > ncolstep) then
+    print*,' ERROR in column assignment for '//trim(tag)
+    icolumn = ncolstep ! screw up last entry, but don't seg fault
+ endif
+
  imaxcolumnread = max(imaxcolumnread,icolumn)
 
 end function assign_column
@@ -1192,11 +1197,13 @@ subroutine get_rho_from_h(i1,i2,ih,ipmass,irho,required,npartoftype,massoftype,h
           npartoftype(itype) = npartoftype(itype) - 1
           npartoftype(itypemap_unknown_phantom) = npartoftype(itypemap_unknown_phantom) + 1
           if (required(irho)) dat(k,irho) = pmassi*(hfact/abs(hi))**3
+          iphase(k) = -1
        else ! dead particles
           npartoftype(itype) = npartoftype(itype) - 1
           npartoftype(itypemap_unknown_phantom) = npartoftype(itypemap_unknown_phantom) + 1
           nkilled = nkilled + 1
           if (required(irho)) dat(k,irho) = 0.
+          iphase(k) = -2
        endif
    enddo
 else
@@ -1311,6 +1318,41 @@ subroutine check_iphase_matches_npartoftype(i1,i2,iphase,npartoftypei)
 
 end subroutine check_iphase_matches_npartoftype
 
+!------------------------------------------------------------
+! allocate and reallocate the iphase array as needed
+!------------------------------------------------------------
+subroutine allocate_iphase(iphase,nmax,phantomdump,gotbinary)
+ integer*1, allocatable, intent(inout) :: iphase(:)
+ integer, intent(in) :: nmax
+ logical, intent(in) :: phantomdump,gotbinary
+ integer*1, allocatable :: iphase_old(:)
+ integer :: ncopy
+
+ if (allocated(iphase)) then
+    iphase_old = iphase ! allocates memory for iphase_old as well
+    deallocate(iphase)
+ endif
+ allocate(iphase(nmax))
+
+ if (phantomdump) then
+    iphase(:) = 1
+ else
+    iphase(:) = 0
+ endif
+
+ if (gotbinary) then
+    iphase(nmax-1) = -3
+    iphase(nmax)   = -3
+ endif
+
+ if (allocated(iphase_old)) then
+    ncopy = min(size(iphase_old),size(iphase))
+    iphase(1:ncopy) = iphase_old(1:ncopy)
+    deallocate(iphase_old)
+ endif
+
+end subroutine allocate_iphase
+
 end module sphNGread
 
 !----------------------------------------------------------------------
@@ -1372,8 +1414,8 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
  real(sing_prec) :: r4
  real, dimension(:,:), allocatable :: dattemp2
  real, dimension(maxinblock) :: dummyreal
- real :: hfact,omega,Xfrac,Yfrac,xHIi,xHIIi,xHeIi,xHeIIi,xHeIIIi,nei,Tcut
- logical :: skip_corrupted_block_3,get_temperature,get_ionfrac
+ real :: hfact,omega,Xfrac,Yfrac,xHIi,xHIIi,xHeIi,xHeIIi,xHeIIIi,nei
+ logical :: skip_corrupted_block_3,get_temperature,get_ionfrac,need_to_allocate_iphase
  character(len=lentag) :: tagsreal(maxinblock), tagtmp
 
  integer, parameter :: splash_max_iversion = 1
@@ -1671,7 +1713,7 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
 
        !--force read of h and rho if dustfrac is required
        if (ndustarrays > 0 .and. any(required(nhydroarrays+1:nhydroarrays+ndustarrays))) then
-          !print*,' dustfrac in columns ',nhydroarrays+1,nhydroarrays+ndustarrays,' required = ',required(nhydroarrays+1)
+          if (debug) print*,' dustfrac in columns ',nhydroarrays+1,nhydroarrays+ndustarrays,' required = ',required(nhydroarrays+1)
           required(irho) = .true.
           required(ih) = .true.
        endif
@@ -1718,6 +1760,7 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
        enddo
     endif
 
+    need_to_allocate_iphase = (npart_max > maxpart) .or. .not.allocated(iphase)
     if (npart_max > maxpart .or. j > maxstep .or. ncolumns > maxcol) then
        if (lowmemorymode) then
           call alloc(max(npart_max+2,maxpart),j,ilastrequired)
@@ -1747,20 +1790,11 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
        !--stop reading file here if no columns required
        !
        if (ilastrequired==0) exit over_MPIblocks
-
-       if (allocated(iphase)) deallocate(iphase)
-       allocate(iphase(npart_max+2))
-       if (phantomdump) then
-          iphase(:) = 1
-       else
-          iphase(:) = 0
-       endif
-
-       if (gotbinary) then
-          iphase(npart-1) = -3
-          iphase(npart)   = -3
-       endif
     endif
+!
+!--allocate memory for iphase array now that gotbinary is known
+!
+    if (need_to_allocate_iphase) call allocate_iphase(iphase,max(npart_max+2,maxpart),phantomdump,gotbinary)
 !
 !--Arrays
 !
@@ -1813,7 +1847,6 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
              gotiphase = .true.
              if (tagged) read(iunit,end=33,iostat=ierr) ! skip tags
              read(iunit,end=33,iostat=ierr) iphase(i1:i2)
-             call check_iphase_matches_npartoftype(i1,i2,iphase,npartoftype(:,j))
              !--skip remaining integer arrays
              nskip = nint1(iarr) - 1 + nint2(iarr) + nint4(iarr) + nint8(iarr)
           endif
@@ -1853,6 +1886,7 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
                 print "(a)",'ERROR: not enough arrays written for sink particles in phantom dump'
                 nskip = nreal(iarr)
              else
+                if (debug) print*,'DEBUG: denoting ',npart,'->',npart+isize(iarr),' as sink particles'
                 iphase(npart+1:npart+isize(iarr)) = -3
                 ilocvx = nreal(iarr)-2 ! velocity is always last 3 numbers for phantom sinks
                 if (doubleprec) then
@@ -1991,7 +2025,7 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
           do i=1,nreal(iarr)
              tagtmp = ''
              if (tagged) read(iunit,end=33,iostat=ierr) tagtmp
-             icolumn = assign_column(tagtmp,iarr,i,6,imaxcolumnread,idustarr)
+             icolumn = assign_column(tagtmp,iarr,i,6,imaxcolumnread,idustarr,ncolstep)
              if (tagged) tagarr(icolumn) = tagtmp
              if (debug)  print*,' reading real to col:',icolumn,' tag = ',trim(tagtmp), ' required = ',required(icolumn)
              if (required(icolumn)) then
@@ -2046,7 +2080,7 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
           do i=1,nreal4(iarr)
              tagtmp = ''
              if (tagged) read(iunit,end=33,iostat=ierr) tagtmp
-             icolumn = assign_column(tagtmp,iarr,i,4,imaxcolumnread,idustarr)
+             icolumn = assign_column(tagtmp,iarr,i,4,imaxcolumnread,idustarr,ncolstep)
              if (debug) print*,'reading real4 to col:',icolumn,' tag = ',trim(tagtmp),' required = ',required(icolumn)
              if (tagged) tagarr(icolumn) = tagtmp
 
@@ -2073,7 +2107,7 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
           do i=1,nreal8(iarr)
              tagtmp = ''
              if (tagged) read(iunit,end=33,iostat=ierr) tagtmp
-             icolumn = assign_column(tagtmp,iarr,i,8,imaxcolumnread,idustarr)
+             icolumn = assign_column(tagtmp,iarr,i,8,imaxcolumnread,idustarr,ncolstep)
              if (debug) print*,'reading real8 to col:',icolumn,' tag = ',trim(tagtmp),' required = ',required(icolumn)
              if (tagged) tagarr(icolumn) = tagtmp
              if (required(icolumn)) then
@@ -2179,7 +2213,6 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
  !--set flag to indicate that only part of this file has been read
  if (.not.all(required(1:ncolstep))) ipartialread = .true.
 
-
  nptmassi = 0
  nunknown = 0
  ngas  = 0
@@ -2187,6 +2220,10 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
  nstar = 0
  !--can only do this loop if we have read the iphase array
  iphasealloc: if (allocated(iphase)) then
+!
+!--sanity check the iphase array
+!
+    if (gotiphase) call check_iphase_matches_npartoftype(1,npart,iphase,npartoftype(:,j))
 !
 !--translate iphase into particle types (mixed type storage)
 !
@@ -2326,6 +2363,7 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
     npartoftype(5,j) = nstar
     npartoftype(6,j) = nunknown
  else
+    if (debug) print*,' DEBUG: nunknown = ',nunknown
     npartoftype(1,j) = npartoftype(1,j) - nunknown
     npartoftype(itypemap_unknown_phantom,j) = npartoftype(itypemap_unknown_phantom,j) + nunknown
  endif